pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chat-enabled-dataroom"
version = "0.1.0"
description = "A chat-enabled dataroom for real estate documents with AI-powered Q&A using LangGraph agents"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["real-estate", "document-analysis", "ai", "chat", "rag", "langgraph", "agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core AI/ML dependencies
    "google-generativeai>=0.3.0",
    "langchain>=0.1.0",
    "langchain-google-genai>=0.1.0",
    "langchain-community>=0.1.0",
    "langgraph>=0.1.0",
    
    # Document processing
    "unstructured[pdf]>=0.10.0",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    
    # Vector database
    "chromadb>=0.4.0",
    
    # Web framework
    "gradio>=4.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
dataroom = "src.dataroom.main:main"

[project.urls]
Homepage = "https://github.com/yourusername/chat-enabled-dataroom"
Repository = "https://github.com/yourusername/chat-enabled-dataroom.git"
Issues = "https://github.com/yourusername/chat-enabled-dataroom/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["dataroom"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true